Reverted client_unix. Introduced separate jump host file.
